Upgraded sprite rendering code.

+++ src/SpaceInvaders/GameOverUIPanel.cpp
@@ -8,13 +8,13 @@ GameOverUIPanel::GameOverUIPanel(Vector2i res) :
 	UIPanel(res,
 		(res.x / 10) * 3,
 		res.y / 2, // 50% of the resolution from the top
-		(res.x / 10) * 3, // as wide as 1/3 of the resolution
-		res.y / 6, // and as tall as 1/6 of the resolution
+		static_cast<float>((res.x / 10) * 3), // as wide as 1/3 of the resolution
+		static_cast<float>(res.y / 6), // and as tall as 1/6 of the resolution
 		50, 255, 255, 255) // a, r, g, b
 {
-	m_ButtonWidth = res.x / 20;
-	m_ButtonHeight = res.y / 20;
-	m_ButtonPadding = res.x / 100;
+	m_ButtonWidth = static_cast<float>(res.x / 20);
+	m_ButtonHeight = static_cast<float>(res.y / 20);
+	m_ButtonPadding = static_cast<float>(res.x / 100);
 	m_Text.setFillColor(Color(0, 255, 0, 255));// Green
 	m_Text.setString("GAME OVER!");
 	m_Font.loadFromFile("fonts/Roboto-Bold.ttf");
@@ -26,9 +26,9 @@ GameOverUIPanel::GameOverUIPanel(Vector2i res) :
 
 void GameOverUIPanel::initialiseButtons()
 {
-	addButton(m_ButtonPadding, m_ButtonPadding, m_ButtonWidth, m_ButtonHeight, 0, 255, 0, "Play");
+	addButton(m_ButtonPadding, m_ButtonPadding, static_cast<int>(m_ButtonWidth), static_cast<int>(m_ButtonHeight), 0, 255, 0, "Play");
 
-	addButton(m_ButtonWidth + (m_ButtonPadding * 2), m_ButtonPadding, m_ButtonWidth, m_ButtonHeight, 255, 0, 0, "Home");
+	addButton(m_ButtonWidth + (m_ButtonPadding * 2), m_ButtonPadding, static_cast<int>(m_ButtonWidth), static_cast<int>(m_ButtonHeight), 255, 0, 0, "Home");
 }
 
 void GameOverUIPanel::draw(RenderWindow& window)
